backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
boto3==1.34.131
pydantic==2.7.4
qrcode==7.4.2
Pillow==10.3.0